+++ src/hooks/LevelInfoLayer.cpp
@@ -6,6 +6,8 @@ using namespace geode::prelude;
 class $modify(DIBLevelInfoLayer, LevelInfoLayer) {
     struct Fields {
         EventListener<web::WebTask> m_listener;
+        bool m_createDemon;
+        CCMenuItemSpriteExtra* m_demonInfoButton;
     };
 
     static void onModify(auto& self) {
@@ -17,18 +19,19 @@ class $modify(DIBLevelInfoLayer, LevelInfoLayer) {
 
         if (level->m_stars.value() < 10) return true;
 
-        auto createDemon = true;
+        auto f = m_fields.self();
+        f->m_createDemon = true;
 
-        if (getChildByID("grd-difficulty") || !Mod::get()->getSettingValue<bool>("enable-difficulties")) createDemon = false;
+        if (getChildByID("grd-difficulty") || !Mod::get()->getSettingValue<bool>("enable-difficulties")) f->m_createDemon = false;
 
         auto gddpDifficulty = getChildByID("gddp-difficulty");
-        if (gddpDifficulty && !Mod::get()->getSettingValue<bool>("gddp-integration-override")) createDemon = false;
+        if (gddpDifficulty && !Mod::get()->getSettingValue<bool>("gddp-integration-override")) f->m_createDemon = false;
         else if (gddpDifficulty) gddpDifficulty->setVisible(false);
 
         auto levelID = level->m_levelID.value();
         auto demon = DemonsInBetween::demonForLevel(levelID, false);
         if (demon.id != 0 && demon.difficulty != 0) {
-            createUI(demon, createDemon);
+            createUI(demon);
             return true;
         }
 
@@ -38,8 +41,8 @@ class $modify(DIBLevelInfoLayer, LevelInfoLayer) {
         ), 3);
         m_difficultySprite->setOpacity(0);
 
-        DemonsInBetween::loadDemonForLevel(std::move(m_fields->m_listener), levelID, false, [this, createDemon](LadderDemon& demon) {
-            createUI(demon, createDemon);
+        DemonsInBetween::loadDemonForLevel(std::move(m_fields->m_listener), levelID, false, [this](LadderDemon& demon) {
+            createUI(demon);
             release();
         }, [this] { retain(); });
 
@@ -83,7 +86,7 @@ class $modify(DIBLevelInfoLayer, LevelInfoLayer) {
             case 6: originalDifficulty = "Extreme Demon"; break;
         }
 
-        FLAlertLayer::create("Demon Info", fmt::format(
+        createQuickPopup("Demon Info", fmt::format(
             "<cy>{}</c>\n"
             "<cg>Tier</c>: {}\n"
             "<cl>Enjoyment</c>: {}\n"
@@ -94,21 +97,31 @@ class $modify(DIBLevelInfoLayer, LevelInfoLayer) {
             demon.enjoyment >= 0.0 ? fmt::format("{}", demon.enjoyment) : "N/A",
             difficulty,
             originalDifficulty
-        ), "OK")->show();
+        ), "OK", "Refresh", [this](auto, bool btn2) {
+            if (!btn2) return;
+            auto levelID = m_level->m_levelID.value();
+            DemonsInBetween::LEVELS_LOADED.erase(levelID);
+            DemonsInBetween::loadDemonForLevel(std::move(m_fields->m_listener), levelID, false, [this](LadderDemon& demon) {
+                createUI(demon);
+                release();
+            }, [this] { retain(); });
+        });
     }
 
-    void createUI(LadderDemon const& demon, bool createDemon) {
-        auto demonInfoButton = CCMenuItemSpriteExtra::create(
+    void createUI(LadderDemon const& demon) {
+        auto f = m_fields.self();
+        if (f->m_demonInfoButton) f->m_demonInfoButton->removeFromParent();
+
+        f->m_demonInfoButton = CCMenuItemSpriteExtra::create(
             CircleButtonSprite::createWithSpriteFrameName(fmt::format("DIB_{:02d}_001.png"_spr, demon.difficulty).c_str()),
             this, menu_selector(DIBLevelInfoLayer::onDemonInfo)
         );
-        demonInfoButton->setID("demon-info-button"_spr);
-
+        f->m_demonInfoButton->setID("demon-info-button"_spr);
         auto leftSideMenu = getChildByID("left-side-menu");
-        leftSideMenu->addChild(demonInfoButton);
+        leftSideMenu->addChild(f->m_demonInfoButton);
         leftSideMenu->updateLayout();
 
-        if (!createDemon) return;
+        if (!f->m_createDemon) return;
 
         if (auto betweenDifficultySprite = static_cast<CCSprite*>(getChildByID("between-difficulty-sprite"_spr))) {
             betweenDifficultySprite->setDisplayFrame(